serve/requirements.txt
fastapi
uvicorn[standard]
torch==2.7.0
torchvision==0.22.0
pillow
numpy
python-multipart